move to_query and use counter for all variable names
